+++ pytransform3d/uncertainty.py
@@ -298,74 +298,6 @@ def concat_dependent_uncertain_transforms(
     return mean_A2C, cov_A2C
 
 
-def to_ellipse(cov, factor=1.0):
-    """Compute error ellipse.
-
-    An error ellipse shows equiprobable points of a 2D Gaussian distribution.
-
-    Parameters
-    ----------
-    cov : array-like, shape (2, 2)
-        Covariance of the Gaussian distribution.
-
-    factor : float, optional (default: 1)
-        One means standard deviation.
-
-    Returns
-    -------
-    angle : float
-        Rotation angle of the ellipse.
-
-    width : float
-        Width of the ellipse (semi axis, not diameter).
-
-    height : float
-        Height of the ellipse (semi axis, not diameter).
-    """
-    from scipy import linalg
-    vals, vecs = linalg.eigh(cov)
-    order = vals.argsort()[::-1]
-    vals, vecs = vals[order], vecs[:, order]
-    angle = np.arctan2(*vecs[:, 0][::-1])
-    width, height = factor * np.sqrt(vals)
-    return angle, width, height
-
-
-def plot_error_ellipse(ax, mean, cov, color=None, alpha=0.25,
-                       factors=np.linspace(0.25, 2.0, 8)):
-    """Plot error ellipse of MVN.
-
-    Parameters
-    ----------
-    ax : axis
-        Matplotlib axis.
-
-    mean : array-like, shape (2,)
-        Mean of the Gaussian distribution.
-
-    cov : array-like, shape (2, 2)
-        Covariance of the Gaussian distribution.
-
-    color : str, optional (default: None)
-        Color in which the ellipse should be plotted
-
-    alpha : float, optional (default: 0.25)
-        Alpha value for ellipse
-
-    factors : array, optional (default: np.linspace(0.25, 2.0, 8))
-        Multiples of the standard deviations that should be plotted.
-    """
-    from matplotlib.patches import Ellipse
-    for factor in factors:
-        angle, width, height = to_ellipse(cov, factor)
-        ell = Ellipse(xy=mean, width=2.0 * width, height=2.0 * height,
-                      angle=np.degrees(angle))
-        ell.set_alpha(alpha)
-        if color is not None:
-            ell.set_color(color)
-        ax.add_artist(ell)
-
-
 def to_ellipsoid(mean, cov):
     """Compute error ellipsoid.
